refactor viewmodels, navigationservice, caching service
moved from presentation to application layer
