+++ SatisfactorySmartHub/SatisfactorySmartHub.Application/Installer/DependencyInjection.cs
@@ -21,6 +21,9 @@ public static class DependencyInjection
     public static IServiceCollection AddApplicationServices(this IServiceCollection services)
     {
         services.RegisterServices();
+        services.AddWindowModels();
+        services.AddViewModels();
+        services.AddNavigation();
         return services;
     }
 }
