+++ SatisfactorySmartHub/SatisfactorySmartHub.Application/Interfaces/Application/Services/ICachingService.cs
@@ -1,12 +1,7 @@
 ﻿using SatisfactorySmartHub.Domain.Models;
-using System;
-using System.Collections.Generic;
 using System.ComponentModel;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace SatisfactorySmartHub.Presentation.Interfaces.Services;
+namespace SatisfactorySmartHub.Application.Interfaces.Application.Services;
 public interface ICachingService : INotifyPropertyChanging, INotifyPropertyChanged
 {
     CorporationModel ActiveCorporation { get; set; }
