+++ SatisfactorySmartHub/SatisfactorySmartHub.Application/SatisfactorySmartHub.Application.csproj
@@ -11,6 +11,7 @@
 	</ItemGroup>
 	
 	<ItemGroup>
+	  <PackageReference Include="CommunityToolkit.Mvvm" Version="8.2.2" />
 	  <PackageReference Include="Microsoft.Extensions.Hosting" Version="8.0.0" />
 	</ItemGroup>
