+++ SatisfactorySmartHub/SatisfactorySmartHub.Application/Services/CachingService.cs
@@ -1,0 +1,15 @@
+﻿using SatisfactorySmartHub.Application.Interfaces.Application.Services;
+using SatisfactorySmartHub.Domain.Common;
+using SatisfactorySmartHub.Domain.Models;
+
+namespace SatisfactorySmartHub.Application.Services;
+internal class CachingService : ObservableObjectBase, ICachingService
+{
+    private CorporationModel? _activeCorporationModel;
+
+    public CorporationModel? ActiveCorporation
+    {
+        get => _activeCorporationModel;
+        set => SetProperty(ref _activeCorporationModel, value);
+    }
+}
