+++ SatisfactorySmartHub/SatisfactorySmartHub.Application/ViewModels/AdminViewModel.cs
@@ -1,29 +1,25 @@
 ﻿using CommunityToolkit.Mvvm.Input;
-using Microsoft.Win32;
 using SatisfactorySmartHub.Application.Interfaces.Application.Services;
 using SatisfactorySmartHub.Application.Interfaces.Infrastructure.Persistence;
+using SatisfactorySmartHub.Application.ViewModels.Base;
 using SatisfactorySmartHub.Domain.Models;
-using SatisfactorySmartHub.Presentation.Common;
-using SatisfactorySmartHub.Presentation.Common.Interfaces;
-using SatisfactorySmartHub.Presentation.Interfaces.Services;
-using SatisfactorySmartHub.Presentation.ViewModels.Base;
 
-namespace SatisfactorySmartHub.Presentation.ViewModels;
+namespace SatisfactorySmartHub.Application.ViewModels;
 
 public sealed class AdminViewModel : ViewModelBase
 {
-    private readonly INavigationHelper _navigationHelper;
+    private readonly INavigationService _navigationHelper;
     private readonly ICorporationService _corporationService;
     private readonly ICachingService _cachingService;
     private readonly IUserDataService _userOptionsHelper;
 
     private IRelayCommand? _corporationCommand;
     private IRelayCommand? _saveCommand;
-    private IRelayCommand? _exportCommand;
     private string _saveHint = string.Empty;
+    private bool _exportPossible = false;
 
     public AdminViewModel(
-        INavigationHelper navigationHelper,
+        INavigationService navigationHelper,
         ICorporationService corporationService,
         ICachingService cachingService,
         IUserDataService userOptionsHelper)
@@ -34,18 +30,35 @@ public sealed class AdminViewModel : ViewModelBase
         _userOptionsHelper = userOptionsHelper;
 
         CorporationCommand.Execute(this);
+
+        if (_cachingService.ActiveCorporation is not null)
+            ExportPossible = true;
+
     }
 
-    public INavigationHelper NavigationHelper => _navigationHelper;
+    public INavigationService NavigationHelper => _navigationHelper;
     public string SaveHint
     {
         get => _saveHint;
         set => SetProperty(ref _saveHint, value);
     }
+    public bool ExportPossible
+    {
+        get => _exportPossible;
+        set => SetProperty(ref _exportPossible, value);
+    }
+    public string ExportName => _cachingService.ActiveCorporation.Name ?? string.Empty;
 
     public IRelayCommand CorporationCommand => _corporationCommand ??= new RelayCommand(NavigationHelper.NavigateMainWindowTo<CorporationViewModel>);
     public IRelayCommand SaveCommand => _saveCommand ?? new RelayCommand(new Action(SaveCorporation));
-    public IRelayCommand ExportCommand => _exportCommand ?? new RelayCommand(new Action(ExportCorporation));
+
+
+    public void ExportCorporation(string filepath)
+    {
+        CorporationModel exportCorporation = _cachingService.ActiveCorporation;
+        _corporationService.ExportCorporation(exportCorporation, filepath);
+        SaveHint = "Export erfolgreich.";
+    }
 
     private void SaveCorporation()
     {
@@ -61,33 +74,6 @@ public sealed class AdminViewModel : ViewModelBase
         SaveHint = "Speichern erfolgreich.";
     }
 
-    private void ExportCorporation()
-    {
-        if (_cachingService.ActiveCorporation == null)
-        {
-            SaveHint = "Kein Konzern zum exportieren geladen.";
-            return;
-        }
-
-        string filepath;
-
-        var saveFileDialog = new SaveFileDialog()
-        {
-            Filter = "json-Datei | *.json",
-            DefaultExt = "json",
-            FileName = _cachingService.ActiveCorporation.Name,
-        };
-
-        if (saveFileDialog.ShowDialog() == true)
-        {
-            filepath = saveFileDialog.FileName;
-        }
-        else
-            return;
-
-        CorporationModel exportCorporation = _cachingService.ActiveCorporation;
-        _corporationService.ExportCorporation(exportCorporation, filepath);
-        SaveHint = "Export erfolgreich.";
-    }
+ 
 
 }
