+++ SatisfactorySmartHub/SatisfactorySmartHub.Application/ViewModels/Base/ViewModelBase.cs
@@ -1,0 +1,11 @@
+﻿using SatisfactorySmartHub.Domain.Common;
+using System.ComponentModel;
+
+namespace SatisfactorySmartHub.Application.ViewModels.Base;
+
+
+/// <summary>
+/// The view model base class.
+/// </summary>
+public abstract class ViewModelBase : ObservableObjectBase, INotifyPropertyChanging, INotifyPropertyChanged
+{ }
