--- SatisfactorySmartHub/SatisfactorySmartHub.Presentation/Common/Interfaces/ViewModels/Base/IViewModelBase.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace SatisfactorySmartHub.Presentation.Common.Interfaces.ViewModels.Base;
-
-/// <summary>
-/// The view model base interface.
-/// </summary>
-public interface IViewModelBase : INotifyPropertyChanging, INotifyPropertyChanged
-{ }
