+++ SatisfactorySmartHub/SatisfactorySmartHub.Presentation/Utilities/ViewDataTemplates.xaml
@@ -2,18 +2,18 @@
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:v="clr-namespace:SatisfactorySmartHub.Presentation.Views"
-    xmlns:vm="clr-namespace:SatisfactorySmartHub.Presentation.ViewModels">
+    xmlns:vm="clr-namespace:SatisfactorySmartHub.Application.ViewModels;assembly=SatisfactorySmartHub.Application">
 
     <DataTemplate DataType="{x:Type TypeName=vm:HubViewModel}">
-        <v:HubView />
+        <v:HubView/>
     </DataTemplate>
 
     <DataTemplate DataType="{x:Type TypeName=vm:AdminViewModel}">
-        <v:AdminView />
+        <v:AdminView/>
     </DataTemplate>
 
     <DataTemplate DataType="{x:Type TypeName=vm:CorporationViewModel}">
-        <v:CorporationView />
+        <v:CorporationView/>
     </DataTemplate>
 
 </ResourceDictionary>
