--- SatisfactorySmartHub/SatisfactorySmartHub.Presentation/ViewModels/Base/ViewModelBase.cs
@@ -1,16 +1,0 @@
-﻿using SatisfactorySmartHub.Domain.Common;
-using SatisfactorySmartHub.Presentation.Common.Interfaces.ViewModels.Base;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace SatisfactorySmartHub.Presentation.ViewModels.Base;
-
-
-/// <summary>
-/// The view model base class.
-/// </summary>
-public abstract class ViewModelBase : ObservableObjectBase, IViewModelBase
-{ }
