+++ SatisfactorySmartHub/SatisfactorySmartHub.Presentation/Views/AdminView.xaml.cs
@@ -1,4 +1,6 @@
-﻿using System;
+﻿using Microsoft.Win32;
+using SatisfactorySmartHub.Application.ViewModels;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -24,5 +26,29 @@ namespace SatisfactorySmartHub.Presentation.Views
         {
             InitializeComponent();
         }
+
+        private void ExportButton_Click(object sender, RoutedEventArgs e)
+        {
+            AdminViewModel dc = (AdminViewModel)DataContext;
+            string filepath;
+
+
+            var saveFileDialog = new SaveFileDialog()
+            {
+                Filter = "json-Datei | *.json",
+                DefaultExt = "json",
+                FileName = dc.ExportName,
+            };
+
+            if (saveFileDialog.ShowDialog() == true)
+            {
+                filepath = saveFileDialog.FileName;
+            }
+            else
+                return;
+
+            dc.ExportCorporation(filepath);
+        }
+
     }
 }
