+++ SatisfactorySmartHub/SatisfactorySmartHub.Presentation/Views/CorporationView.xaml
@@ -5,7 +5,7 @@
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:local="clr-namespace:SatisfactorySmartHub.Presentation.Views"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    xmlns:vm="clr-namespace:SatisfactorySmartHub.Presentation.ViewModels;assembly=SatisfactorySmartHub.Presentation"
+    xmlns:vm="clr-namespace:SatisfactorySmartHub.Application.ViewModels;assembly=SatisfactorySmartHub.Application"
     d:DataContext="{d:DesignInstance Type=vm:CorporationViewModel}"
     d:DesignHeight="450"
     d:DesignWidth="800"
