+++ SatisfactorySmartHub/SatisfactorySmartHub.Presentation/Views/HubView.xaml
@@ -5,7 +5,7 @@
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:local="clr-namespace:SatisfactorySmartHub.Presentation.Views"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    xmlns:vm="clr-namespace:SatisfactorySmartHub.Presentation.ViewModels;assembly=SatisfactorySmartHub.Presentation"
+    xmlns:vm="clr-namespace:SatisfactorySmartHub.Application.ViewModels;assembly=SatisfactorySmartHub.Application"
     d:DataContext="{d:DesignInstance Type=vm:HubViewModel}"
     d:DesignHeight="450"
     d:DesignWidth="800"
@@ -91,7 +91,7 @@
                     VerticalAlignment="Top"
                     Background="{StaticResource grey4}"
                     BorderThickness="0"
-                    Command="{Binding ImportCorporationCommand}"
+                    Click="LoadButton_Click"
                     Content="import"
                     Foreground="{StaticResource FontWhite}" />
