+++ SatisfactorySmartHub/SatisfactorySmartHub.Presentation/Views/HubView.xaml.cs
@@ -1,18 +1,6 @@
-﻿using SatisfactorySmartHub.Presentation.ViewModels;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows;
+﻿using Microsoft.Win32;
+using SatisfactorySmartHub.Application.ViewModels;
 using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Navigation;
-using System.Windows.Shapes;
 
 namespace SatisfactorySmartHub.Presentation.Views;
 
@@ -24,5 +12,25 @@ public partial class HubView : UserControl
     public HubView()
     {
         InitializeComponent();
+    }
+
+    private void LoadButton_Click(object sender, System.Windows.RoutedEventArgs e)
+    {
+        string filepath;
+        var openFileDialog = new OpenFileDialog()
+        {
+            Filter = "json-Datei | *.json",
+            DefaultExt = "json",
+        };
+
+        if (openFileDialog.ShowDialog() == true)
+        {
+            filepath = openFileDialog.FileName;
+        }
+        else
+            return;
+
+        HubViewModel dc = (HubViewModel)DataContext;
+        dc.ImportCorporation(filepath);
     }
 }
