+++ SatisfactorySmartHub/SatisfactorySmartHub.Presentation/Windows/MainWindow.xaml.cs
@@ -1,17 +1,5 @@
-﻿using SatisfactorySmartHub.Presentation.WindowModels;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using SatisfactorySmartHub.Application.WindowModels;
 using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Shapes;
 
 namespace SatisfactorySmartHub.Presentation.Windows
 {
